+++ packages/react/src/lib/godaddy/checkout-queries.ts
@@ -495,6 +495,12 @@ export const DraftOrderQuery = graphql(`
                   name
                   ratePercentage
               }
+              taxes {
+                  amount {
+                      value
+                  }
+                  included
+              }
               totals {
                   discountTotal {
                       currencyCode